+++ src/main/java/prismatic/shards/stellarity/mixin/extend_classes/EndSpikeMixin.java
@@ -3,7 +3,9 @@ package prismatic.shards.stellarity.mixin.extend_classes;
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.mojang.serialization.Codec;
 import net.minecraft.world.level.levelgen.feature.EndSpikeFeature;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import prismatic.shards.stellarity.interface_injection.ExtEndSpike;
@@ -11,11 +13,17 @@ import prismatic.shards.stellarity.util.CodecExtensionHelper;
 
 @Mixin(EndSpikeFeature.EndSpike.class)
 public class EndSpikeMixin implements ExtEndSpike {
+	@Shadow
+	@Final
+	public static Codec<EndSpikeFeature.EndSpike> CODEC;
 	@Unique
 	private boolean altar;
 
 	@Unique
 	private boolean cryingObsidianTops;
+
+	@Unique
+	private boolean hasCrystal;
 
 	@Override
 	public boolean stellarity$hasAltar() {
@@ -33,6 +41,16 @@ public class EndSpikeMixin implements ExtEndSpike {
 	}
 
 	@Override
+	public boolean stellarity$hasCrystal() {
+		return hasCrystal;
+	}
+
+	@Override
+	public void stellarity$setHasCrystal(boolean hasCrystal) {
+		this.hasCrystal = hasCrystal;
+	}
+
+	@Override
 	public void stellarity$setCryingObsidianTops(boolean cryingObsidianTops) {
 		this.cryingObsidianTops = cryingObsidianTops;
 	}
@@ -46,8 +64,9 @@ public class EndSpikeMixin implements ExtEndSpike {
 	)
 	private static Codec<EndSpikeFeature.EndSpike> more(Codec<EndSpikeFeature.EndSpike> original) {
 		return CodecExtensionHelper.buildExtensionCodec(original, (instance, wrapper) -> instance.group(wrapper,
-				Codec.BOOL.fieldOf("stellarity:altar").orElse(false).forGetter(ExtEndSpike::stellarity$hasAltar),
-				Codec.BOOL.fieldOf("stellarity:crying_obsidian_tops").orElse(false).forGetter(ExtEndSpike::stellarity$hasCryingObsidianTops)
+				Codec.BOOL.optionalFieldOf("stellarity:altar", false).forGetter(ExtEndSpike::stellarity$hasAltar),
+				Codec.BOOL.optionalFieldOf("stellarity:crying_obsidian_tops", false).forGetter(ExtEndSpike::stellarity$hasCryingObsidianTops),
+				Codec.BOOL.optionalFieldOf("stellarity:has_crystal", true).forGetter(ExtEndSpike::stellarity$hasCrystal)
 			).apply(instance, ExtEndSpike::apply),
 			ExtEndSpike::applyDefaults
 		);